requirements.txt
streamlit
Pillow
torch
transformers
